refactor test4-repair
